+++ src/tests/App.test.js
@@ -1,11 +1,11 @@
 import React from "react";
 import { shallow } from "enzyme";
 
-import GameContainer from "./../containers/GameContainer";
+import Game from "./../containers/Game";
 import App from "../App";
 
 describe(`App`, () => {
-  it("should render a `GameContainer`", () => {
-    expect(shallow(<App />).find(GameContainer).length).toBe(1);
+  it("should render a `Game`", () => {
+    expect(shallow(<App />).find(Game).length).toBe(1);
   });
 });
